+++ CHANGELOG.md
@@ -12,7 +12,7 @@ PDBTools.jl Changelog
 
 Version 3.5.0-DEV
 -------------
-- ![FEATURE][badge-feature] `sasa` function to compute solvent accessible surface area.
+- ![FEATURE][badge-feature] `atomic_sasa` and `sasa` functions to compute solvent accessible surface area.
 - ![FEATURE][badge-feature] Add `vdw_radius` as element property, and `element_vdw_radius` function to fetch them.
 - ![INFO][badge-info] The element symbol is returned as a `InlineStrings.String3`. 
 - ![INFO][badge-info] Masses are returned as `Float32` values instead of `Float64`.